changes on memories.ts, new oauth, login authentication on backend

+++ nlw Spacetime/server/src/routes/memories.ts
@@ -4,8 +4,15 @@ import { z } from 'zod'
 import { FastifyInstance } from "fastify";
 
 export async function memoriesRoutes(app: FastifyInstance){
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => { // antes de executar handler(funçoes abaixo) ele verifica se o usario esta autenticado com JWT token
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -20,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance){
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id : z.string().uuid(),
     })
@@ -33,6 +40,10 @@ export async function memoriesRoutes(app: FastifyInstance){
         id,
       },
     })
+
+    if(!memory.isPublic && memory.userId != request.user.sub){
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -51,13 +62,14 @@ export async function memoriesRoutes(app: FastifyInstance){
         content,
         coverUrl,
         isPublic,
-        userId: 'aaf37696-19dc-4a5e-a551-059495dfb272'
+        userId: request.user.sub,
       }
     })
 
     return memory
   })
-  app.put('/memories/:id', async (request) => {
+
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id : z.string().uuid(),
     })
@@ -72,7 +84,17 @@ export async function memoriesRoutes(app: FastifyInstance){
 
     const { content, coverUrl, isPublic} = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    }) 
+    
+    if(memory.userId != request.user.sub){
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -86,12 +108,22 @@ export async function memoriesRoutes(app: FastifyInstance){
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id : z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    }) 
+    
+    if(memory.userId != request.user.sub){
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
